MARTIAN-250 #comment Possible fix for spurious stage_defs error
When restarting a pipestance still kept in memory, split_has_run
and join_has_run would not be reset to false. Therefore, the split
job would often not run and not generate a stage_defs file. Martian
runtime would, however, think the job was run and generate the
stage_defs error.