Allow JSON path expressions in WHERE and ORDER BY clauses.
This is a major refactor to JSONExtract2 so that the 'WhereAble' and 'sortkey' components may
now contain JSON path expressions.  We accept these in the same format that we used for the
select expression.

The critical differences are:
 * Inside a where clause, any substring that looks like "/A/B/C" will
interpreted as a JSON path expression.

 * The type translation from objects returned from postgres works
differently. We now get objects with their real type and do
transformation to keep unrepresentable floating point types from ruining
things.