Stabalize column sorts; add latest-version filter.
This enforces a stable sort order on the columns in "metrics" and
"basic" table views.

This also adds a "latest-version" filter. When enabled, LIGO trims
pipestances. For each LENA ID present in the results set, it only
shows the results from the highest versioned pipeline. It also restricts
results to successful pipestances from "mario".

This feature only works of the page contains every possible result
from the query. If a page has been paginated, it will fail.