Stop wasting my life on tab-versus-space problems.
We all know that Eight shall be the number thou shalt tab and the number
of the tabbing shall be eight. Nine shalt thou not tab neither count thou
FOUR, excepting that thou then proceed to eight.

However, lack of standardization on this matter is causing stupid bugs
when tabs and spaces get mixed in jade files.  This diff
 1) Removes tabs from LIGO's jade templates
 2) Changes ligo to crash if it ever sees a tab in a template

This also addresses MARSOC-161.