Add caching for deserialized json metadata.
Only cache things which are deserialized to json, which is generally a
smaller set, and invalidate the cache on write or update.

This should address the issue the that was observed running the new VDJ
pipelines, where a stage with thousands of chunks depended on a stage
with thousands of outputs, resulting in millions of file accesses when
trying to resolve argument bindings.