Switch the qstat check to also handle running jobs.
Instead of just checking to make sure queued jobs are actually queued
and then letting them time out, check both queued and running jobs to
make sure they're actually queued or running.  If they are not, fail the
job immediately (after the check completes.  The check is still
throttled to every 5 minutes).

Complicating this change was the fact that the queue check happens in a
separate goroutine, to ensure that it doesn't block progress of the main
loop.  To make this function reliably, checked_reset (used by autorestart)
was made more atomic, to ensure that we can detect the case of checking
if a job is still running, but then discovering an errors file for it
and autorestarting it before the query returns.

Also made output for martian_test more verbose for file content
mismatch.

Also fixed the sge_queue script to properly treat jobs in the 'Eqw'
state as errored, as opposed to queued.