Fix a race condition in the qstat check. (#229)
When qstat comes back with a report that the job is not running, rather
than immediately failing out the job instead wait until the next time
the node's state is refreshed from the journal.

Also, when taking action based on the output of qstat, dump the stderr
of the qstat wrapper script to the martian log.  Put the raw output of
qstat in the stderr of the wrapper script.