Add a configurable grace period for finished jobs.
After qstat sees a job has finished, the job's metadata files may take
some time before they are visible to mrp because of NFS delays.

The default timeout is set to 1 hour if nothing is found in the config.
For SGE the default will be 50 minutes on customer systems (so that it's
slightly less than the heartbeat timeout) and 40 seconds for Hydra.