Check heartbeat time against last metadata refresh. (#233)
In some cases, if the run loop gets stalled (e.g. by an nfs outage or
the computer going into standby at the wrong moment), the heartbeat can
be checked a long time after the metadata was updated.  Rather than
failing out the stage at that time, compare the heartbeat to the last
time we checked for a heartbeat update.