Switch martian-public to a submodule. (#253)
Because the cs tarball depends on redstone, which we are not open
sourcing, we still build with mostly the old makefile.  Adapters, tools,
and jobmanagers are copied up to the root tree, since sake-strip removes
src.  Martian web is also built in the root so that editor and mrv get
added back in.