Update public submodule.
Brings in update to mrp for always-on UI port (autoselected by OS if not
provided).

Since this removes teh martini submodules from martian-public, they
needed to be added back to the internal submodules.

Also, adding an authentication token to the graph template means
updating the GraphPage object everywhere that has a copy of this.
Breaking that code out into an importable module would be a good plan,
except we're planning on just using mrp for everything in the future.